+++ src/modules/api/spoonacular/spoonacular.api.utils.ts
@@ -9,6 +9,7 @@ export function proceedRecipesToMeals(recipes: SpoonacularRecipe[]): RatedMeal[]
             id: recipe.id.toString(),
             imgUrl: recipe.image,
             ingredients: [...proceedIngredients(recipe.usedIngredients), ...proceedIngredients(recipe.missedIngredients)],
+            missingCount: recipe.missedIngredients.length,
             relevance: recipe.usedIngredients.length / (recipe.usedIngredients.length + recipe.missedIngredients.length),
             title: recipe.title,
             provider: 'spoonacular',