+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.50.0] - 2024-09-16
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `missingCount` property for `RatedMeal`.
+
+### Fixed
+- [Marek Kurańda](https://github.com/mjkuranda): Unit and E2E tests.
+
 ## [1.49.0] - 2024-08-31
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): `readyInMinutes` property for meals from MongoDB.